Revise mui menu & items
