+++ src/components/Mui/Menu/Menu.scss
@@ -1,3 +1,29 @@
 .mui-menu-paper.mui-menu-paper {
   background-color: var(--main-lighter-color);
 }
+
+.mui-menu-item.mui-menu-item {
+  @include padding-y(8px);
+  color: var(--text-color);
+  font-size: inherit;
+
+  > div {
+    @include flex(center, space-between);
+
+    flex: 1 1 auto;
+    max-width: 100%;
+
+    .text {
+      @include text-overflow-ellipsis();
+
+      flex: 1 1 auto;
+      padding-right: 25px;
+    }
+  }
+
+  &:hover {
+    [data-theme^='dark'] & {
+      background-color: var(--main-light-color);
+    }
+  }
+}
