+++ src/components/Mui/Menu/index.ts
@@ -1,5 +1,6 @@
 import './Menu.scss';
 
 export * from './Menu';
+export * from './MenuItem';
 export * from './useMuiMenu';
 export { Menu as default } from './Menu';
