--- src/components/Mui/MenuItem/MenuItem.scss
@@ -1,25 +1,0 @@
-.mui-menu-item.mui-menu-item {
-  @include padding-y(8px);
-  color: var(--text-color);
-  font-size: inherit;
-
-  > div {
-    @include flex(center, space-between);
-
-    flex: 1 1 auto;
-    max-width: 100%;
-
-    .text {
-      @include text-overflow-ellipsis();
-
-      flex: 1 1 auto;
-      padding-right: 25px;
-    }
-  }
-
-  &:hover {
-    [data-theme^='dark'] & {
-      background-color: var(--main-light-color);
-    }
-  }
-}
