--- src/components/Mui/MenuItem/index.ts
@@ -1,4 +1,0 @@
-import './MenuItem.scss';
-
-export * from './MenuItem';
-export { useMenuItem as default } from './MenuItem';
